mvoe CustomToggle and CustomMenu to their own files for reuse purposes
