+++ src/Components/CustomToggle.jsx
@@ -1,0 +1,18 @@
+import React from "react";
+// The forwardRef is important!!
+// Dropdown needs access to the DOM node in order to position the Menu
+const CustomToggle = React.forwardRef(({ children, onClick }, ref) => (
+    <a
+        href="."
+        ref={ref}
+        onClick={(e) => {
+            e.preventDefault();
+            onClick(e);
+        }}
+    >
+        {children}
+        &#x25bc;
+    </a>
+));
+
+export default CustomToggle;
